change in nav bar - clinets and services list and view

+++ database/migrations/2018_03_04_194633_create_clients_table.php
@@ -22,7 +22,7 @@ class CreateClientsTable extends Migration
             //client's email (required) and two clients cannot have the same email
             $table->string('email')->nullable(false)->unique();
             //client's phone number (required) and two clients cannot have the same phone number
-            $table->smallInteger('phone_number')->unique();
+            $table->Integer('phone_number')->unique();
             //client's adress
             $table->string('address')->nullable();
